Adds custom title tag/working on makeLinks()
Currently trying to edit nav1 in custom_config.php so that makeLinks(nav1)  in header.php link to proper path.  When on news page, clicking "News" again stays on the same page, and when clicking "ADD News", it links to the create page.  When on create page, clicking on "ADD News" again stays on the same page.  This is an improvement from before, as clicking on current page's link was adding new path.  In the process of fixing links so that while on ANY page, navigation links are able to link to any page desired.  Committing these changes because as of now, this is the best working version.  This version also adds custom title tags.